requirements.txt
amqp==1.4.7
anyjson==0.3.3
awscli==1.7.42
Babel==2.1.1
backports-abc==0.4
beautifulsoup4==4.4.1
billiard==3.3.0.21
boto3==1.1.1
botocore==1.1.12
celery==3.1.19
colorama==0.3.3
Django==1.8
django-braces==1.8.1
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-loaderio==0.3.0
django-oauth-toolkit==0.9.0
djangorestframework==3.1.3
docutils==0.12
flower==0.8.3
futures==2.2.0
google==1.7
greenlet==0.4.9
gunicorn==19.4.5
httplib2==0.9.2
jmespath==0.7.1
keen==0.3.20
kombu==3.0.29
livestreamer==1.12.2
Markdown==2.6.2
oauthlib==1.0.3
Padding==0.4
pubnub==3.7.4
pyasn1==0.1.8
pycrypto==2.6.1
PyMySQL==0.6.6
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
redis==2.10.5
requests==2.8.0
rsa==3.1.4
six==1.9.0
sqlparse==0.1.18
tornado==4.3
virtualenv==13.1.2
webapp2==2.5.2
